requirements.txt
aiogram==3.0.0
asyncio
logging
json
python-dotenv
